// frontend/web/src/services/API/proserve-wb-provisioning-api/models/GetProvisionedProductResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AvailableVersionDistribution } from './AvailableVersionDistribution';
import {
    AvailableVersionDistributionFromJSON,
    AvailableVersionDistributionFromJSONTyped,
    AvailableVersionDistributionToJSON,
} from './AvailableVersionDistribution';
import type { ProvisionedProduct } from './ProvisionedProduct';
import {
    ProvisionedProductFromJSON,
    ProvisionedProductFromJSONTyped,
    ProvisionedProductToJSON,
} from './ProvisionedProduct';

/**
 * 
 * @export
 * @interface GetProvisionedProductResponse
 */
export interface GetProvisionedProductResponse {
    /**
     * 
     * @type {ProvisionedProduct}
     * @memberof GetProvisionedProductResponse
     */
    provisionedProduct?: ProvisionedProduct;
    /**
     * 
     * @type {AvailableVersionDistribution}
     * @memberof GetProvisionedProductResponse
     */
    versionMetadata?: AvailableVersionDistribution;
}

/**
 * Check if a given object implements the GetProvisionedProductResponse interface.
 */
export function instanceOfGetProvisionedProductResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetProvisionedProductResponseFromJSON(json: any): GetProvisionedProductResponse {
    return GetProvisionedProductResponseFromJSONTyped(json, false);
}

export function GetProvisionedProductResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProvisionedProductResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'provisionedProduct': !exists(json, 'provisionedProduct') ? undefined : ProvisionedProductFromJSON(json['provisionedProduct']),
        'versionMetadata': !exists(json, 'versionMetadata') ? undefined : AvailableVersionDistributionFromJSON(json['versionMetadata']),
    };
}

export function GetProvisionedProductResponseToJSON(value?: GetProvisionedProductResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'provisionedProduct': ProvisionedProductToJSON(value.provisionedProduct),
        'versionMetadata': AvailableVersionDistributionToJSON(value.versionMetadata),
    };
}